MDEV-21469 multi-engine part.
Refines the parent patch to handle recovery of XA involving
two or more engines.

Multiple engine case requires to recognize partly prepared XA.
For that xa_recovery_member::in_engine_prepare is turned into a counter;
the counter is compared against the # of engine recovery participants
to decide how/if to prepare and complete.